frontend/requirements.txt
streamlit==1.28.1
requests==2.31.0
pandas==2.1.2
Pillow==10.1.0